// frontend/src/generated-api/models/ChatMessage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Transcendence backend
 * The Transcendence API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { User } from './User';
import {
    UserFromJSON,
    UserFromJSONTyped,
    UserToJSON,
    UserToJSONTyped,
} from './User';
import type { ChatRoom } from './ChatRoom';
import {
    ChatRoomFromJSON,
    ChatRoomFromJSONTyped,
    ChatRoomToJSON,
    ChatRoomToJSONTyped,
} from './ChatRoom';

/**
 * 
 * @export
 * @interface ChatMessage
 */
export interface ChatMessage {
    /**
     * 
     * @type {number}
     * @memberof ChatMessage
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    content: string;
    /**
     * 
     * @type {Date}
     * @memberof ChatMessage
     */
    sentTime: Date;
    /**
     * 
     * @type {number}
     * @memberof ChatMessage
     */
    userId: number;
    /**
     * 
     * @type {number}
     * @memberof ChatMessage
     */
    chatRoomId: number;
    /**
     * 
     * @type {User}
     * @memberof ChatMessage
     */
    user: User;
    /**
     * 
     * @type {ChatRoom}
     * @memberof ChatMessage
     */
    chatRoom: ChatRoom;
}

/**
 * Check if a given object implements the ChatMessage interface.
 */
export function instanceOfChatMessage(value: object): value is ChatMessage {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    if (!('sentTime' in value) || value['sentTime'] === undefined) return false;
    if (!('userId' in value) || value['userId'] === undefined) return false;
    if (!('chatRoomId' in value) || value['chatRoomId'] === undefined) return false;
    if (!('user' in value) || value['user'] === undefined) return false;
    if (!('chatRoom' in value) || value['chatRoom'] === undefined) return false;
    return true;
}

export function ChatMessageFromJSON(json: any): ChatMessage {
    return ChatMessageFromJSONTyped(json, false);
}

export function ChatMessageFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatMessage {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'content': json['content'],
        'sentTime': (new Date(json['sent_time'])),
        'userId': json['user_id'],
        'chatRoomId': json['chat_room_id'],
        'user': UserFromJSON(json['user']),
        'chatRoom': ChatRoomFromJSON(json['chatRoom']),
    };
}

export function ChatMessageToJSON(json: any): ChatMessage {
    return ChatMessageToJSONTyped(json, false);
}

export function ChatMessageToJSONTyped(value?: ChatMessage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'content': value['content'],
        'sent_time': ((value['sentTime']).toISOString()),
        'user_id': value['userId'],
        'chat_room_id': value['chatRoomId'],
        'user': UserToJSON(value['user']),
        'chatRoom': ChatRoomToJSON(value['chatRoom']),
    };
}